+++ src/main/java/quentin/core/Stone.java
@@ -1,0 +1,15 @@
+package quentin.core;
+
+public enum Stone {
+    WHITE,
+    BLACK,
+    NONE;
+
+    public Stone getOppositeColor() {
+        return switch (this) {
+            case BLACK -> WHITE;
+            case WHITE -> BLACK;
+            case NONE -> NONE;
+        };
+    }
+}
