+++ src/main/java/quentin/exceptions/IllegalMoveException.java
@@ -1,0 +1,10 @@
+package quentin.exceptions;
+
+import quentin.core.Position;
+
+public class IllegalMoveException extends Exception {
+    public IllegalMoveException(Position position){
+        super("Player cannot put a stone diagonally adjacent to another stone of the same colour this" + position +
+                " without a colour alike orthogonally adjacent.");
+    }
+}
