+++ src/main/java/quentin/exceptions/InvalidFirstPlayerException.java
@@ -1,0 +1,8 @@
+package quentin.exceptions;
+
+public class InvalidFirstPlayerException extends Exception {
+
+    public InvalidFirstPlayerException(){
+        super("Black player should play first");
+    }
+}
