+++ src/main/java/quentin/exceptions/OccupiedPositionException.java
@@ -1,0 +1,11 @@
+package quentin.exceptions;
+
+import quentin.core.Position;
+
+public class OccupiedPositionException extends Exception {
+
+    public OccupiedPositionException(Position position){
+        super(position + "is already occupied.");
+    }
+}
+
