+++ src/main/java/quentin/exceptions/OutOfBoardException.java
@@ -1,0 +1,10 @@
+package quentin.exceptions;
+
+public class OutOfBoardException  extends  Exception{
+    public OutOfBoardException(String error)
+    {
+        super(error);
+    }
+
+}
+
