+++ src/main/java/quentin/exceptions/PieException.java
@@ -1,0 +1,8 @@
+package quentin.exceptions;
+
+public class PieException extends  Exception{
+    public PieException(String error){
+        super(error);
+    }
+
+}
