+++ src/main/java/quentin/exceptions/RepeatedPlayException.java
@@ -1,0 +1,7 @@
+package quentin.exceptions;
+
+public class RepeatedPlayException extends Exception {
+    public RepeatedPlayException(){
+        super("A player cannot play twice in a row.");
+    }
+}
