--- src/main/java/sdmExam/core/Stone.java
@@ -1,15 +1,0 @@
-package sdmExam.core;
-
-public enum Stone {
-    WHITE,
-    BLACK,
-    NONE;
-
-    public Stone getOppositeColor() {
-        return switch (this) {
-            case BLACK -> WHITE;
-            case WHITE -> BLACK;
-            case NONE -> NONE;
-        };
-    }
-}
