--- src/main/java/sdmExam/exceptions/IllegalMoveException.java
@@ -1,10 +1,0 @@
-package sdmExam.exceptions;
-
-import sdmExam.core.Position;
-
-public class IllegalMoveException extends Exception {
-    public IllegalMoveException(Position position){
-        super("Player cannot put a stone diagonally adjacent to another stone of the same colour this" + position +
-                " without a colour alike orthogonally adjacent.");
-    }
-}
