--- src/main/java/sdmExam/exceptions/InvalidFirstPlayerException.java
@@ -1,8 +1,0 @@
-package sdmExam.exceptions;
-
-public class InvalidFirstPlayerException extends Exception {
-
-    public InvalidFirstPlayerException(){
-        super("Black player should play first");
-    }
-}
