--- src/main/java/sdmExam/exceptions/OccupiedPositionException.java
@@ -1,11 +1,0 @@
-package sdmExam.exceptions;
-
-import sdmExam.core.Position;
-
-public class OccupiedPositionException extends Exception {
-
-    public OccupiedPositionException(Position position){
-        super(position + "is already occupied.");
-    }
-}
-
