--- src/main/java/sdmExam/exceptions/OutOfBoardException.java
@@ -1,10 +1,0 @@
-package sdmExam.exceptions;
-
-public class OutOfBoardException  extends  Exception{
-    public OutOfBoardException(String error)
-    {
-        super(error);
-    }
-
-}
-
