--- src/main/java/sdmExam/exceptions/PieException.java
@@ -1,8 +1,0 @@
-package sdmExam.exceptions;
-
-public class PieException extends  Exception{
-    public PieException(String error){
-        super(error);
-    }
-
-}
