--- src/main/java/sdmExam/exceptions/RepeatedPlayException.java
@@ -1,7 +1,0 @@
-package sdmExam.exceptions;
-
-public class RepeatedPlayException extends Exception {
-    public RepeatedPlayException(){
-        super("A player cannot play twice in a row.");
-    }
-}
